fix: serializing by constructor call

+++ libs/core/std/include/rtbot/std/StandardDeviation.h
@@ -13,7 +13,7 @@ template <class T, class V>
 struct StandardDeviation : public Buffer<T, V> {
   StandardDeviation() = default;
 
-  StandardDeviation(string const &id_, unsigned int n_) : Buffer<T, V>(id_, n_) {}
+  StandardDeviation(string const &id_, size_t n_) : Buffer<T, V>(id_, n_) {}
 
   string typeName() const override { return "StandardDeviation"; }
